+++ messaging/messaging-core/src/main/java/uk/gov/justice/services/messaging/logging/HttpMessageLoggerHelper.java
@@ -1,18 +1,6 @@
 package uk.gov.justice.services.messaging.logging;
 
-import static java.lang.String.join;
-import static javax.json.Json.createObjectBuilder;
-import static javax.ws.rs.core.HttpHeaders.ACCEPT;
-import static javax.ws.rs.core.HttpHeaders.CONTENT_TYPE;
-import static uk.gov.justice.services.common.http.HeaderConstants.CLIENT_CORRELATION_ID;
-import static uk.gov.justice.services.common.http.HeaderConstants.ID;
-import static uk.gov.justice.services.common.http.HeaderConstants.NAME;
-import static uk.gov.justice.services.common.http.HeaderConstants.SESSION_ID;
-import static uk.gov.justice.services.common.http.HeaderConstants.USER_ID;
-import static uk.gov.justice.services.common.log.LoggerConstants.METADATA;
-
 import javax.inject.Inject;
-import javax.json.JsonObjectBuilder;
 import javax.ws.rs.core.HttpHeaders;
 import javax.ws.rs.core.MultivaluedMap;